+++ src/lib/descriptors/Material/RawShaderMaterialDescriptor.js
@@ -5,18 +5,6 @@ class RawShaderMaterialDescriptor extends ShaderMaterialDescriptor {
   construct(props) {
     const materialDescription = this.getMaterialDescription(props);
 
-    if (props.hasOwnProperty('uniforms')) {
-      materialDescription.uniforms = props.uniforms;
-    }
-
-    if (props.hasOwnProperty('vertexShader')) {
-      materialDescription.vertexShader = props.vertexShader;
-    }
-
-    if (props.hasOwnProperty('fragmentShader')) {
-      materialDescription.fragmentShader = props.fragmentShader;
-    }
-
     return new THREE.RawShaderMaterial(materialDescription);
   }
 }